requirements.txt
google-generativeai
chromadb
SpeechRecognition
pyttsx3
pyaudio
